requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
python-dotenv==1.0.0
email-validator==2.0.0
imaplib2==3.6
openai==0.28.1
scikit-learn==1.3.1
python-dateutil==2.8.2
requests==2.31.0
Werkzeug==2.3.7
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
google-api-python-client==2.100.0
PyJWT==2.8.0
